MachineVision/all-requirements.txt
awscli==1.29.54
blinker==1.6.2
boto3==1.28.54
botocore==1.31.54
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
colorama==0.4.4
contourpy==1.1.1
cycler==0.11.0
docker==3.4.0
docker-pycreds==0.4.0
docutils==0.16
filelock==3.12.4
Flask==2.3.3
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.37
idna==3.4
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jmespath==1.0.1
keyring==8.7
keyrings.alt==5.0.0
kiwisolver==1.4.5
lit==17.0.1
MarkupSafe==2.1.3
matplotlib==3.8.0
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.7.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
smmap==5.0.1
sympy==1.12
tabulate==0.9.0
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.187
urllib3==1.26.16
voluptuous==0.13.1
websocket-client==1.6.3
Werkzeug==2.3.7